refactor: refactor inventory system to support spread-based free draw credits and introduce decimal precision for card enhancements

+++ Backend/src/TarotNow.Application/Features/Gamification/Dtos/CardStatsDto.cs
@@ -10,11 +10,11 @@ public class CardStatsDto
     public int Level { get; set; }
 
     // Chỉ số tấn công.
-    public int Atk { get; set; }
+    public decimal Atk { get; set; }
 
     // Chỉ số phòng thủ.
-    public int Def { get; set; }
+    public decimal Def { get; set; }
 
     // Tổng sức mạnh quy đổi từ ATK + DEF.
-    public int TotalPower => Atk + Def;
+    public decimal TotalPower => Atk + Def;
 }
